refactoring: extract part of functionality in StatusPanel class #7199
